requirements.txt
pytest-mock==1.12.1
pytest-cov==2.8.1
moto==1.3.14
pytest==5.3.5
bandit==1.6.2
pylint==2.6.0
tox==3.20.1
safety==1.10.1